+++ compare-frontend/src/app/modules/wizard/converters/eqts/3_exposure.ts
@@ -1,6 +1,7 @@
 import { ScenarioName, RiesgosScenarioState, RiesgosProductResolved } from "src/app/state/state";
 import { Converter } from "../../converter.service";
 import { WizardComposite } from "../../wizard.service";
+import { LegendComponent } from "../../tabComponents/legends/legendComponents/legend/legend.component";
 
 
 export class Exposure implements Converter {
@@ -23,10 +24,20 @@ export class Exposure implements Converter {
                 currentValue: currentValue
             }],
             step: step,
-            // legend: {
-            //     component: fdsafs
-            //     args: {}
-            // }
+            legend: () => ({
+                component: LegendComponent,
+                args: {
+                    text: 'exposureLegend',
+                    entries: [{
+                    text: 'Exposure',
+                    color: '#fdfdfd'
+                    }, {
+                    text: 'NoData',
+                    color: '#c1c1c1'
+                    }],
+                    height: 50
+                }
+            })
         }
     }